correct to importance word in expression search

+++ main.cpp
@@ -56,5 +56,5 @@ TEST_CASE("Check if found terms in collection") {
 
     REQUIRE(search.getDocumentsFound().at(0).getName() == "doc3");
     REQUIRE(search.getDocumentsFound().at(1).getName() == "doc1");
-    REQUIRE(search.getWordImportance().find("apartamento")->second == log(3/2));
+    REQUIRE(search.getWordImportance().find("apartamento")->second == log((float) 3/2));
 }
